[vid 65] Started adding rule-based access control Started building out rule-based access control for the app.
NOTE: There's an odd issue where my roles are being stored in an object
inside of an array. This differs from what is seen in the course, and
I'm unsure whether this is something I've set up wonky, or if there's
some other explanation. I've opened up a thread in the course Slack
and am waiting for input from the team/community.
For now, I have resolved the issue by simply accessing the object
inside the array. But, I may need to revisit this and address an
underlying issue at some point in the future.